⚡️ perf: refactor the responsive layout to improve performance (#1541)
