+++ src/app/chat/(desktop)/index.tsx
@@ -4,7 +4,8 @@ import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
-import ResponsiveIndex from '@/components/ResponsiveIndex';
+import ResponsiveContainer from '@/components/ResponsiveContainer';
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import PageTitle from '../features/PageTitle';
 import ChatHeader from './features/ChatHeader';
@@ -12,10 +13,13 @@ import Conversation from './features/Conversation';
 import SideBar from './features/SideBar';
 import Layout from './layout.desktop';
 
-const Mobile: FC = dynamic(() => import('../(mobile)'), { ssr: false }) as FC;
+const Mobile: FC = dynamic(() => import('../(mobile)'), {
+  loading: MobileSwitchLoading,
+  ssr: false,
+}) as FC;
 
 const DesktopPage = memo(() => (
-  <ResponsiveIndex Mobile={Mobile}>
+  <ResponsiveContainer Mobile={Mobile}>
     <Layout>
       <PageTitle />
       <ChatHeader />
@@ -24,6 +28,6 @@ const DesktopPage = memo(() => (
         <SideBar />
       </Flexbox>
     </Layout>
-  </ResponsiveIndex>
+  </ResponsiveContainer>
 ));
 export default DesktopPage;
