+++ src/app/chat/settings/(desktop)/index.tsx
@@ -4,20 +4,24 @@ import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
 import SessionHydration from '@/app/chat/components/SessionHydration';
-import ResponsiveIndex from '@/components/ResponsiveIndex';
+import ResponsiveContainer from '@/components/ResponsiveContainer';
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import EditPage from '../features/EditPage';
 import Layout from './layout.desktop';
 
-const Mobile: FC = dynamic(() => import('../(mobile)'), { ssr: false }) as FC;
+const Mobile: FC = dynamic(() => import('../(mobile)'), {
+  loading: MobileSwitchLoading,
+  ssr: false,
+}) as FC;
 
 const ChatSettings = memo(() => (
   <>
-    <ResponsiveIndex Mobile={Mobile}>
+    <ResponsiveContainer Mobile={Mobile}>
       <Layout>
         <EditPage />
       </Layout>
-    </ResponsiveIndex>
+    </ResponsiveContainer>
     <SessionHydration />
   </>
 ));
