+++ src/app/market/(desktop)/index.tsx
@@ -4,19 +4,23 @@ import { SpotlightCard, SpotlightCardProps } from '@lobehub/ui';
 import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
-import AgentCard from '@/app/market/features/AgentCard';
-import ResponsiveIndex from '@/components/ResponsiveIndex';
+import ResponsiveContainer from '@/components/ResponsiveContainer';
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
+import AgentCard from '../features/AgentCard';
 import Index from '../index';
 import Layout from './layout.desktop';
 
-const Mobile: FC = dynamic(() => import('../(mobile)'), { ssr: false }) as FC;
+const Mobile: FC = dynamic(() => import('../(mobile)'), {
+  loading: MobileSwitchLoading,
+  ssr: false,
+}) as FC;
 
 export default memo(() => (
-  <ResponsiveIndex Mobile={Mobile}>
+  <ResponsiveContainer Mobile={Mobile}>
     <Layout>
       <Index />
       <AgentCard CardRender={SpotlightCard as FC<SpotlightCardProps>} />
     </Layout>
-  </ResponsiveIndex>
+  </ResponsiveContainer>
 ));
