+++ src/app/settings/(desktop)/index.tsx
@@ -3,19 +3,23 @@
 import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
-import ResponsiveIndex from '@/components/ResponsiveIndex';
+import ResponsiveContainer from '@/components/ResponsiveContainer';
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 import { SettingsTabs } from '@/store/global/initialState';
 
 import Common from '../common';
 import { SettingsCommonProps } from '../common/Common';
 import DesktopLayout from './layout.desktop';
 
-const Mobile: FC = dynamic(() => import('../(mobile)'), { ssr: false }) as FC;
+const Mobile: FC = dynamic(() => import('../(mobile)'), {
+  loading: MobileSwitchLoading,
+  ssr: false,
+}) as FC;
 
 export default memo<SettingsCommonProps>((props) => (
-  <ResponsiveIndex Mobile={Mobile}>
+  <ResponsiveContainer Mobile={Mobile}>
     <DesktopLayout activeTab={SettingsTabs.Common}>
       <Common {...props} />
     </DesktopLayout>
-  </ResponsiveIndex>
+  </ResponsiveContainer>
 ));
