+++ src/app/welcome/(desktop)/index.tsx
@@ -3,19 +3,23 @@
 import dynamic from 'next/dynamic';
 import { FC, memo } from 'react';
 
-import ResponsiveIndex from '@/components/ResponsiveIndex';
+import ResponsiveContainer from '@/components/ResponsiveContainer';
+import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import Footer from './features/Footer';
 import Showcase from './features/Showcase';
 import Layout from './layout.desktop';
 
-const Mobile: FC = dynamic(() => import('../(mobile)'), { ssr: false }) as FC;
+const Mobile: FC = dynamic(() => import('../(mobile)'), {
+  loading: MobileSwitchLoading,
+  ssr: false,
+}) as FC;
 
 export default memo(() => (
-  <ResponsiveIndex Mobile={Mobile}>
+  <ResponsiveContainer Mobile={Mobile}>
     <Layout>
       <Showcase />
       <Footer />
     </Layout>
-  </ResponsiveIndex>
+  </ResponsiveContainer>
 ));
